% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlots.r
\name{calcYlim}
\alias{calcYlim}
\title{Calculate upper YLim for an absolute expression boxplot.}
\usage{
calcYlim(.data, plotBy, padding = 0.2)
}
\arguments{
\item{.data}{A grouped tibble}

\item{plotBy}{The column that will be plotted. Must be an expresison of form
expr(!!sym(gene)).}

\item{padding}{Amount of padding to be applied. Default = 0.2}
}
\value{
An integer vector of form c(0, upper_ylim).
}
\description{
It is advisable to cut off extreme outliers from RNA expression
data so the Y-Limit stays within reasonable levels to show the interesting
data. This function sets the Y-Limit to the whisker height of the highest group,
plus the padding value (0.2 = 20% more than highest col).
}
\examples{
ylim <- calcYlim(filteredCellData, plotBy = expr(!!sym("MYC")))
ylim <- calcYlim(filtredPatientData, plotBy = expr(!!sym("BASP1")), 0.5)
}
