% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCbioData.r
\name{combineMultipleCancerTypes}
\alias{combineMultipleCancerTypes}
\title{Make an expression table for multiple cancer types.}
\usage{
combineMultipleCancerTypes(inputTable, genes, molecularProfile)
}
\arguments{
\item{inputTable}{A 2-column dataframe, colnames have to be "origin" and "studies".}

\item{genes}{Character vector of genes to evaluate}

\item{molecularProfile}{The type of molecular profile to query. For most
patient cohorts it is "rna_seq_v2_mrna", for cell lines "rna_seq_mrna".}
}
\value{
A tibble of RNA expression values, grouped by cancer origin,
one row per sample.
}
\description{
A table of cancer types and their respecitve studies
(as character vector or raw cbioPortal URL) is searched and output in a table
grouped by cancer origin.
}
\examples{
\dontrun{
allPatientData <- combineMultipleCancerTypes(table, c("MTOR", "RPTOR"), "rna_seq_v2_mrna")
}

}
