% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCbioData.r
\name{fetchCbioData}
\alias{fetchCbioData}
\title{Title}
\usage{
fetchCbioData(gene1, gene2)
}
\arguments{
\item{gene1}{"Numerator" gene.}

\item{gene2}{"Denominator" gene.}
}
\value{
Writes tibbles "filteredPatientData", "filteredCellData", and "selectedCellData"
to a global variable.
}
\description{
Title
}
\examples{
fetchCbioData("MYC", "BASP1")
}
