% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCbioData.r
\name{filterByOrigin}
\alias{filterByOrigin}
\title{Filter cBioPortal data by cancer origin.}
\usage{
filterByOrigin(expressionData, filterList, column = origin)
}
\arguments{
\item{expressionData}{A Tibble containing the expression data}

\item{filterList}{A character vector containing the cancer types to filter}

\item{column}{Column name to filter by. Default is "origin".}
}
\value{
Tibble of filtered expression data
}
\description{
Filter cBioPortal data by cancer origin.
}
\examples{
\dontrun{
filterByOrigin(allPatientData, c("Bowel", "Breast", "Melanoma"))
}

}
