% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCbioData.r
\name{getMolecularData}
\alias{getMolecularData}
\title{Get expression data from a single study.}
\usage{
getMolecularData(study, genes, molecularProfile)
}
\arguments{
\item{study}{A studyId string, e.g. "ccle_broad_2019".}

\item{genes}{A character vector of HUGO gene names.}

\item{molecularProfile}{A molecularProfile stub to search for. Usually "rna_seq_v2_mrna" or similar}
}
\value{
A table of expression values as given by cBioPortalData.
}
\description{
gets RNA expression profiles from cBioPortal for the given study and genes
}
\examples{
getMolecularData("ccle_broad_2019", c("MYC", "BASP1"), "rna_seq_mrna")
}
