% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlots.r
\name{plotCellsTable}
\alias{plotCellsTable}
\title{Plot table of expression values and ratios for selected cell lines.}
\usage{
plotCellsTable(.data, gene1, gene2)
}
\arguments{
\item{.data}{Data table containing selected cell line expression data.
Usually you will use selectedCellData generated by fetchCbioData.}

\item{gene1}{Numerator gene.}

\item{gene2}{Denominator gene.}
}
\value{
A ggplot2 object.
}
\description{
Plot table of expression values and ratios for selected cell lines.
}
\examples{
\dontrun{
cellTable <- plotCellsTable (selectedCellData, "MYC", "PHB")
}

}
