% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlots.r
\name{plotPatients}
\alias{plotPatients}
\alias{plotCells}
\title{Plot patient data.}
\usage{
plotPatients(.data, gene1, gene2 = NULL, ylim = NULL)

plotCells(.data, gene1, gene2 = NULL, ylim = NULL)
}
\arguments{
\item{.data}{A tibble containing expression values, generated by
combineMultipleCancerTypes.}

\item{gene1}{"Numerator" gene in ratio plots, or single gene.}

\item{gene2}{Optional denominator gene.}

\item{ylim}{Manual ylim.}
}
\value{
A ggplot2 plot.
}
\description{
Plot patient data.

Plot cell line data.
}
\examples{
patPlot <- plotPatients(filteredPatientData, "MYC", "BASP1")
patPlot1 <- plotPatients(filteredPatientData, "MTOR")
patPlot2 <- plotPatients(filteredPatientData, "RPTOR", ylim = c(0,20000))

}
