% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlots.r
\name{plotRatio}
\alias{plotRatio}
\title{Generate a set of gene ratio plots.}
\usage{
plotRatio(gene1, gene2)
}
\arguments{
\item{gene1}{Numerator gene.}

\item{gene2}{Denominator gene.}
}
\value{
Writes images to OUTPUT_PATH iff DO_WRITE == TRUE, or outputs on
the GUI otherwise.
}
\description{
Takes the objects filteredCellData, filteredPatientData and
selectedCellData, generated by fetchCbioData and outputs the ratio plots.
You want to run fetchCbioData before running this function.
}
\examples{
plotRatio("MYC", "BASP1")
}
