% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{rnaRatio}
\alias{rnaRatio}
\title{rnaRatio}
\usage{
rnaRatio(
  gene1,
  gene2,
  patientInput,
  patientFilter = "",
  cellFilter = "",
  selectedCellFilter = "",
  forceReload = FALSE
)
}
\arguments{
\item{gene1}{The "numerator" gene.}

\item{gene2}{The "denominator" gene,}

\item{patientInput}{A 2-column table with colnames =c("origin", "studies")}

\item{patientFilter}{A character vector of cancer types to include for patients.}

\item{cellFilter}{A character vector of cancer types to include for cells.}

\item{selectedCellFilter}{A character vector of cell lines to specially annotate..}

\item{forceReload}{Force reload of the cBioPortalData.}
}
\value{
Either outputs images or plot to the GUI.
}
\description{
The main function of rnaRatio. You should not need to call
any other functions for normal use.
}
\examples{
\dontrun{
rnaRatio("MYC", "BASP1", patientUrls, patientFilter, cellFilter, selectedCellFilter)
}

}
