% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCbioData.r
\name{toEntrez}
\alias{toEntrez}
\title{toEntrez}
\usage{
toEntrez(gene, translatorTable = translatedGenes)
}
\arguments{
\item{gene}{Either a HUGO gene symbol string or a Entrez ID}

\item{translatorTable}{A translator table generated by translateGenes}
}
\value{
A Entrez Id
}
\description{
Converts a HUGO gene symbol to Entrez using org.Hs.eg.db.
Returns identity if it is already a Entrez symbol.
}
\examples{
toEntrez("MYC")
toEntrez("MTOR")
toEntrez(4609)
toEntrez("4609")

}
