% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCbioData.r
\name{translateGenes}
\alias{translateGenes}
\title{Make a translation table between HUGO and EntrezID.}
\usage{
translateGenes(genelist)
}
\arguments{
\item{genelist}{A numerical or character vector of genes. Accepted are HUGO
and EntrezId.}
}
\value{
A dataframe whith HUGO and EntrezId for each gene
}
\description{
The cBioPortal API internally uses Entrez IDs, which are not
very human-readable. Therefore this function take a given list of HUGO
gene symbols and turns it into a lookup table to use in other functions.
}
\examples{
\dontrun{
tranlatedGenes <- translateGenes(C("MYC", "BASP1", "MTOR", "RPTOR"))
}

}
