% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCbioData.r
\name{fetchCbioData}
\alias{fetchCbioData}
\title{Generate expression table for a set of genes}
\usage{
fetchCbioData(gene1, gene2, ...)
}
\arguments{
\item{gene1}{"Numerator" gene.}

\item{gene2}{"Denominator" gene.}

\item{...}{more optional genes}
}
\value{
Writes tibbles "filteredPatientData", "filteredCellData", and
"selectedCellData" to a global variable.
}
\description{
Generate expression table for a set of genes
}
\examples{
fetchCbioData("MYC", "BASP1")

}
